CustomEvaluableNodeComparator into InterpreterOpcodesTransformations...
... It is needed only there and this removes direct (but not transitive)
dependency of EvaluableNodeTreeFunctions on Interpreter
