+++ src/Amalgam/interpreter/InterpreterOpcodesTransformations.cpp
@@ -19,6 +19,104 @@
 #include <iomanip>
 #include <regex>
 
+class Interpreter;
+
+namespace {
+//used for any operation that must sort different values - for passing in a lambda to run on every operation
+class CustomEvaluableNodeComparator
+{
+public:
+	constexpr CustomEvaluableNodeComparator(Interpreter *_interpreter, EvaluableNode *_function, EvaluableNode *target_list)
+		: interpreter(_interpreter), function(_function), targetList(target_list), hadExecutionSideEffects(false)
+	{ }
+
+	bool operator()(EvaluableNode *a, EvaluableNode *b);
+
+	inline bool DidAnyComparisonHaveExecutionSideEffects()
+	{
+		return hadExecutionSideEffects;
+	}
+
+private:
+	Interpreter *interpreter;
+	EvaluableNode *function;
+	EvaluableNode *targetList;
+	bool hadExecutionSideEffects;
+};
+
+bool CustomEvaluableNodeComparator::operator()(EvaluableNode *a, EvaluableNode *b)
+{
+	//create context with "a" and "b" variables
+	interpreter->PushNewConstructionContext(nullptr, targetList, EvaluableNodeImmediateValueWithType(), a);
+	interpreter->PushNewConstructionContext(nullptr, targetList, EvaluableNodeImmediateValueWithType(), b);
+
+	//compare
+	bool retval = (interpreter->InterpretNodeIntoNumberValue(function) > 0);
+
+	if(interpreter->PopConstructionContextAndGetExecutionSideEffectFlag())
+		hadExecutionSideEffects = true;
+	if(interpreter->PopConstructionContextAndGetExecutionSideEffectFlag())
+		hadExecutionSideEffects = true;
+
+	return retval;
+}
+
+//performs a top-down stable merge on the sub-lists from start_index to middle_index and middle_index to _end_index
+//  from source into destination using cenc
+void CustomEvaluableNodeOrderedChildNodesTopDownMerge(std::vector<EvaluableNode *> &source,
+	size_t start_index, size_t middle_index, size_t end_index, std::vector<EvaluableNode *> &destination, CustomEvaluableNodeComparator &cenc)
+{
+	size_t left_pos = start_index;
+	size_t right_pos = middle_index;
+
+	//for all elements, pull from the appropriate buffer (left or right)
+	for(size_t cur_index = start_index; cur_index < end_index; cur_index++)
+	{
+		//if left_pos has elements left and is less than the right, use it
+		if(left_pos < middle_index && (right_pos >= end_index || cenc(source[left_pos], source[right_pos])))
+		{
+			destination[cur_index] = source[left_pos];
+			left_pos++;
+		}
+		else //the right is less, use that
+		{
+			destination[cur_index] = source[right_pos];
+			right_pos++;
+		}
+	}
+}
+
+//performs a stable merge sort of source (which *will* be modified and is not constant)
+// from start_index to end_index into destination; uses cenc for comparison
+void CustomEvaluableNodeOrderedChildNodesSort(std::vector<EvaluableNode *> &source,
+	size_t start_index, size_t end_index, std::vector<EvaluableNode *> &destination, CustomEvaluableNodeComparator &cenc)
+{
+	//if one element, then sorted
+	if(start_index + 1 >= end_index)
+		return;
+
+	size_t middle_index = (start_index + end_index) / 2;
+
+	//sort left into list
+	CustomEvaluableNodeOrderedChildNodesSort(destination, start_index, middle_index, source, cenc);
+	//sort right into list
+	CustomEvaluableNodeOrderedChildNodesSort(destination, middle_index, end_index, source, cenc);
+
+	//merge buffers back into buffer
+	CustomEvaluableNodeOrderedChildNodesTopDownMerge(source, start_index, middle_index, end_index, destination, cenc);
+}
+
+std::vector<EvaluableNode *> CustomEvaluableNodeOrderedChildNodesSort(std::vector<EvaluableNode *> &list, CustomEvaluableNodeComparator &cenc)
+{
+	//must make two copies of the list to edit, because switch back and forth and there is a chance that an element may be invalid
+	// in either list.  Therefore, can't use the original list in the off chance that something is garbage collected
+	std::vector<EvaluableNode *> list_copy_1(list);
+	std::vector<EvaluableNode *> list_copy_2(list);
+	CustomEvaluableNodeOrderedChildNodesSort(list_copy_1, 0, list.size(), list_copy_2, cenc);
+	return list_copy_2;
+}
+}
+
 EvaluableNodeReference Interpreter::InterpretNode_ENT_REWRITE(EvaluableNode *en, bool immediate_result)
 {
 	auto &ocn = en->GetOrderedChildNodes();
